// src/app/model/solicitacoes.ts
export class Solicitacoes {
   _id: string | null;
    id: number | null;
    empresa: string | null;
    cnpj: string | null;
    plano: string | null;
    tarifa: string | null;
    minutos: string | null;
    vplano: string | null;
    dateAdesao: string | null;
    dateEmissao: string | null;
}